requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.0
scipy>=1.11.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
